requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-flatpickr==1.0.1
django-multiselectfield==0.1.12
djangorestframework==3.11.0
idna==2.8
Pillow==7.1.2
pytz==2019.1
sqlparse==0.3.0
